Clean up PR after initial research
