+++ lib/bgs/services.rb
@@ -22,9 +22,8 @@ require "bgs/services/document"
 require "bgs/services/org"
 require "bgs/services/person"
 require "bgs/services/rating"
-require "bgs/services/rating_comparison"
 require "bgs/services/rating_profile"
-require "bgs/services/rating_web"
+require "bgs/services/vbms_rating"
 require "bgs/services/rating_management"
 require "bgs/services/standard_data"
 require "bgs/services/veteran"
